migrate tier 2 nodes
